# pdoe-backend/PDOE.Admin.API/Controllers/ParametrageController.cs
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using PDOE.Admin.API.Features.GetParametre;
using PDOE.Admin.API.Features.ListParametrage;
using PDOE.Admin.API.Features.UpdateParametre;
using PDOE.Api.Contracts;

namespace PDOE.Admin.API.Controllers;

[ApiController]
[Route("parametrage")]
public class ParametrageController(IMediator mediator) : ControllerBase
{
    // Pas de [Authorize] ici : lecture ouverte à tout utilisateur authentifié (FallbackPolicy, cf. Program.cs) —
    // consommée par DossierCreateComponent (seuils domiciliation/FDI), pas seulement par l'écran d'admin.
    [HttpGet]
    public async Task<ActionResult<List<ParametreMetierResponse>>> ListParametrage(CancellationToken cancellationToken)
    {
        var result = await mediator.Send(new ListParametrageQuery(), cancellationToken);
        return Ok(result);
    }

    [HttpGet("{cle}")]
    public async Task<ActionResult<ParametreMetierResponse>> GetParametre(string cle, CancellationToken cancellationToken)
    {
        var result = await mediator.Send(new GetParametreQuery(cle), cancellationToken);
        return Ok(result);
    }

    [HttpPatch("{cle}")]
    [Authorize(Policy = "AdminDsiri")]
    public async Task<ActionResult<ParametreMetierResponse>> UpdateParametre(
        string cle,
        [FromBody] UpdateParametreRequest request,
        CancellationToken cancellationToken)
    {
        var result = await mediator.Send(new UpdateParametreCommand(cle, request), cancellationToken);
        return Ok(result);
    }
}