# pdoe-backend/PDOE.Shared.Kernel/OpenApi/PDOE_openapi.yaml
openapi: 3.1.0
info:
  title: PDOE API — Plateforme de Digitalisation des Opérations avec l'Extérieur
  version: 5.6.0
  description: |
    API REST de la PDOE pour AfrilandFirstBank CI (AFBCI).

    Orchestration du circuit COMEX en 7 étapes séquentielles :
    Initiation → Validation Gestionnaire → Contrôle COMEX →
    Avis Trésorerie → Exécution SWIFT → Apurement → Archivage.

    **Rejet pour correction** : `POST /workflow/{dossierId}/rejeter` peut
    renvoyer le dossier vers n'importe quelle étape déjà franchie (pas
    seulement l'étape précédente) — `responsableCorrection` désigne cette
    étape cible (`NiveauValidation`), jamais un login ou un texte libre.

    **Authentification** : bind LDAP (Active Directory AFBCI) suivi d'une
    vérification en deux étapes (OTP) — un identifiant/mot de passe valides
    ne délivrent plus directement un JWT, ils déclenchent l'envoi d'un code
    dont la vérification délivre la session réelle (`POST /auth/login` puis
    `POST /auth/otp/verifier`). Le mapping login AD → profil PDOE et le
    statut actif/désactivé d'un compte vivent dans `Utilisateurs`, gérés par
    l'Admin DSIRI (`Auth` ne fait que vérifier l'identité, jamais le profil).

    **Modules** :
    - `Auth` : bind LDAP, vérification OTP, session
    - `Utilisateurs` : annuaire local des profils PDOE (Admin DSIRI)
    - `Portefeuilles` : affectation automatique Gestionnaire ↔ compte client
...
    - `Execution` : bascule SWIFT, déclaration retour
...
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /clients/{numCompte}/verifier-signature:
    get:
      tags: [CBS]
      summary: Vérifier la signature client dans ABS2000
      description: |
        Déclenché automatiquement sur l'événement `onBlur` du champ `numCompte`
        dans le formulaire Angular — avant toute création de dossier en base.
        
        Le comportement dépend du mode configuré dans `ParametrageMetier`
        (clé `MODE_VERIFICATION_SIGNATURE`) :
        
        - **AUTOMATIQUE** : retourne uniquement `signatureExistante: bool`
        - **VISUEL** : retourne `signatureExistante` + `imageSignature` (base64)
        - **LES_DEUX** : retourne les deux
        
        Si le compte est introuvable → 404.
        Si la signature est absente → 422.
        Délai cible de réponse : < 2 secondes.
      operationId: verifierSignatureClient
      parameters:
        - name: numCompte
          in: path
          required: true
          schema:
            type: string
            maxLength: 20
          example: "CI00123456789"
        - name: mode
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ModeVerificationSignature'
          description: Mode de vérification (remplace ParametrageMetier si fourni)
      responses:
        '200':
          description: Compte trouvé et signature valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureVerificationResult'
        '404':
          description: Numéro de compte introuvable dans ABS2000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: COMPTE_INTROUVABLE
                message: "Numéro de compte introuvable dans ABS2000"
        '422':
          description: Compte trouvé mais signature absente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: SIGNATURE_INVALIDE
                message: "Signature client absente dans ABS2000"
  /clients/{numCompte}/valider-signature-visuelle:
    post:
      tags: [CBS]
      summary: Enregistrer la validation visuelle de la signature par l'agent
      description: |
        Appelé après que l'agent d'accueil a coché la case de conformité
        visuelle (modes VISUEL et LES_DEUX). Enregistre les initiales de
        l'agent comme preuve d'audit humaine.
      operationId: validerSignatureVisuelle
      parameters:
        - name: numCompte
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationVisuelleRequest'
      responses:
        '200':
          description: Validation visuelle enregistrée
          content:
            application/json:
              schema:
                type: object
                properties:
                  signatureValidee:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /clients/{numCompte}/solde:
    get:
      tags: [CBS]
      summary: Vérifier le solde du compte client dans ABS2000
      description: |
        Utilisé à l'étape 2 (validation gestionnaire). ABS2000 ne renvoie que le solde brut (lecture seule) —
        suffisant est calculé côté PDOE en comparant soldeDisponible au montant de l'opération du dossier,
        jamais fourni tel quel par ABS2000.
      operationId: getSoldeClient
      parameters:
        - name: numCompte
          in: path
          required: true
          schema:
            type: string
        - name: dossierId
          in: query
          required: true
          description: Dossier dont le montant sert de référence pour le calcul de suffisant.
          schema:
            type: integer
      responses:
        '200':
          description: Solde récupéré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoldeClientResult'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /dossiers/{dossierId}/documents:
    get:
      tags: [Documents]
      summary: Lister les documents d'un dossier
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/DossierIdParam'
        - name: typeDocument
          in: query
          schema:
            $ref: '#/components/schemas/TypeDocument'
      responses:
        '200':
          description: Liste des documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'

    post:
      tags: [Documents]
      summary: Téléverser un document
      description: |
        Calcule le hash SHA-256 côté serveur.
        Stocke le fichier dans le répertoire IIS sécurisé.
        Enregistre les métadonnées en base (hash, chemin, type).
      operationId: uploadDocument
      parameters:
        - $ref: '#/components/parameters/DossierIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '200':
          description: Document enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '413':
          description: Fichier trop volumineux
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dossiers/{dossierId}/documents/{documentId}:
    patch:
      tags: [Documents]
      summary: Valider ou invalider un document
      operationId: updateDocumentStatut
      parameters:
        - $ref: '#/components/parameters/DossierIdParam'
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estValide:
                  type: boolean
      responses:
        '200':
          description: Statut document mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /dossiers/{dossierId}/documents/{documentId}/fichier:
    get:
      tags: [Documents]
      summary: Télécharger/prévisualiser le fichier d'un document
      description: |
        Relit le fichier stocké sur le disque du serveur (Storage:DocumentsRootPath)
        et vérifie son intégrité (SHA-256) avant de le renvoyer.
      operationId: telechargerFichierDocument
      parameters:
        - $ref: '#/components/parameters/DossierIdParam'
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contenu du fichier
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Document introuvable, ou fichier absent du disque
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /parametrage/{cle}:
    get:
      tags: [Parametrage]
      summary: Récupérer un paramètre par clé
      operationId: getParametre
      parameters:
        - name: cle
          in: path
          required: true
          schema:
            type: string
          example: MODE_VERIFICATION_SIGNATURE
      responses:
        '200':
          description: Paramètre trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParametreMetierResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Parametrage]
      summary: Modifier la valeur d'un paramètre (Admin DSIRI ou Super Admin uniquement)
      description: |
        Seuls les paramètres avec `modifiableUI = true` peuvent être modifiés.
        La modification est tracée (UpdatedAt, UpdatedBy).
        Le cache IMemoryCache est invalidé immédiatement.
      operationId: updateParametre
      parameters:
        - name: cle
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParametreRequest'
      responses:
        '200':
          description: Paramètre mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParametreMetierResponse'
        '403':
          description: Paramètre non modifiable via dashboard (modifiableUI = false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Valeur hors plage (valeurMin/valeurMax)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # WORKFLOW-CONFIG — étapes du circuit configurables (Admin DSIRI)
  # ============================================================
  #
  # Les 7 étapes historiques (NiveauValidation) restent des identités
  # stables — Code ne change jamais pour elles. Pas de suppression
  # physique : une étape se désactive (actif = false), jamais ne se
  # supprime, pour les mêmes raisons que /utilisateurs (l'historique
  # EtapesWorkflow doit rester résoluble).
...
  /checklist-config/items:
    get:
      tags: [ChecklistConfig]
      summary: Lister tous les items configurés, actifs et inactifs (Admin DSIRI)
      operationId: listChecklistItemsConfig
      responses:
        '200':
          description: Liste complète, triée par ordre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChecklistItemConfig'

    post:
      tags: [ChecklistConfig]
      summary: Créer un item de checklist (Admin DSIRI)
      operationId: creerChecklistItemConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistItemConfigCreateRequest'
      responses:
        '201':
          description: Item créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItemConfig'
        '400':
          description: libelle manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CHECKLIST_ITEM_LIBELLE_MANQUANT
                message: libelle est requis.
        '403':
          $ref: '#/components/responses/Forbidden'

  # /reordonner AVANT le regex générique /{checklistItemId} — même règle
  # que /workflow-config/etapes/reordonner ci-dessus.
  /checklist-config/items/reordonner:
    patch:
      tags: [ChecklistConfig]
      summary: Réordonnancer l'ensemble des items en une seule opération atomique (Admin DSIRI)
      operationId: reordonnerChecklistItemsConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReordonnerChecklistItemsRequest'
      responses:
        '200':
          description: Nouvel ordre appliqué
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChecklistItemConfig'
        '422':
          description: La liste ne correspond pas exactement aux items existants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CHECKLIST_ORDRE_INVALIDE
                message: La liste fournie ne correspond pas exactement aux items existants.
        '403':
          $ref: '#/components/responses/Forbidden'
  /checklist-config/items/{checklistItemId}:
    patch:
      tags: [ChecklistConfig]
      summary: Modifier le libellé ou l'état actif d'un item (Admin DSIRI)
      operationId: modifierChecklistItemConfig
      parameters:
        - name: checklistItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistItemConfigUpdateRequest'
      responses:
        '200':
          description: Item mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItemConfig'
        '404':
          description: Item introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CHECKLIST_ITEM_INTROUVABLE
                message: Item de checklist introuvable.
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # JOURNAL D'AUDIT (Super Admin uniquement — Admin DSIRI exclu, séparation des tâches)
  # ============================================================
  #
  # Trace tout : authentification, actions admin (utilisateurs, paramétrage,
  # configuration du circuit) et chaque transition de dossier (soumission,
  # validation, rejet, exécution, apurement...). Vue globale cross-dossiers ;
  # GET /workflow/{dossierId}/historique reste l'historique détaillé d'un
  # dossier donné.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT généré par PDOE.Gateway une fois le bind LDAP/AD AFBCI ET la
        vérification OTP réussis (voir tag Auth) — jamais délivré sur le
        seul bind LDAP.
  parameters:
    DossierIdParam:
      name: dossierId
      in: path
      required: true
      schema:
        type: integer
      description: Identifiant unique du dossier (INT IDENTITY)
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Profil insuffisant pour cette action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    QualiteResidence:
      type: string
      enum: [RESIDENT, NON_RESIDENT]
      description: |
        Qualité de la résidence du client au sens de la réglementation
        des changes BCEAO — RESIDENT (personne physique/morale résidente
        de l'UEMOA) ou NON_RESIDENT.
...
    TypeDocument:
      type: string
      enum:
        - PIECE_IDENTITE
        - FACTURE_PROFORMA
        - CONTRAT
        - FDI
        - AUTORISATION_CHANGE
        - BAE
        - D3
        - FACTURE_SOLDEE
        - JUSTIFICATIF_APUREMENT
        - ORDRE_TRANSFERT
        - FORMULAIRE_CHANGE
        - ATTESTATION_CHANGE
        - ATTESTATION_BNC
        - FACTURE_DEFINITIVE
        - AUTRE
...
    ModeVerificationSignature:
      type: string
      enum: [AUTOMATIQUE, VISUEL, LES_DEUX]
      description: |
        Mode de vérification signature ABS2000 :
        - AUTOMATIQUE : contrôle binaire (signature présente ou absente), suffisant seul
        - VISUEL : ABS2000 ne fournit pas l'image — l'agent se rend dans ABS2000 lui-même,
          compare visuellement avec le document papier (ordre de transfert), puis coche
          case + initiales dans PDOE
        - LES_DEUX : automatique ET visuelle obligatoires
...
    UploadDocumentRequest:
      type: object
      required: [fichier, typeDocument]
      properties:
        fichier:
          type: string
          format: binary
        typeDocument:
          $ref: '#/components/schemas/TypeDocument'
        referenceDocument:
          type: string
          maxLength: 100
          nullable: true
          description: "N° de référence du document lui-même (ex : n° de facture, n° d'AC, n° de formulaire de change) — distinct du nom système du fichier uploadé. Pertinent surtout pour FACTURE_PROFORMA/FACTURE_DEFINITIVE/AUTORISATION_CHANGE/FORMULAIRE_CHANGE."
        estObligatoire:
          type: boolean
          default: false
        paiementId:
          type: integer
          nullable: true
    ValidationVisuelleRequest:
      type: object
      required: [initialesAgent]
      properties:
        initialesAgent:
          type: string
          minLength: 1
          maxLength: 10
          description: Initiales de l'agent ayant effectué la vérification visuelle

    # ── Réponses ────────────────────────────────────────────
...
    DocumentResponse:
      type: object
      properties:
        documentId:
          type: integer
        typeDocument:
          $ref: '#/components/schemas/TypeDocument'
        referenceDocument:
          type: string
          nullable: true
        nomFichier:
          type: string
        hashSHA256:
          type: string
          minLength: 64
          maxLength: 64
        tailleFichier:
          type: integer
          format: int64
        estObligatoire:
          type: boolean
        estValide:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
...
    ParametreMetierResponse:
      type: object
      properties:
        parametreId:
          type: integer
        cle:
          type: string
        valeur:
          type: string
        unite:
          type: string
        description:
          type: string
        modifiableUI:
          type: boolean
        valeurMin:
          type: string
          nullable: true
        valeurMax:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    UpdateParametreRequest:
      type: object
      required: [valeur]
      properties:
        valeur:
          type: string
          description: Nouvelle valeur — validée entre valeurMin et valeurMax

    # ── Workflow-config — étapes du circuit configurables ────
...
    ReordonnerChecklistItemsRequest:
      type: object
      required: [ordre]
      properties:
        ordre:
          type: array
          description: |
            Liste COMPLÈTE des checklistItemId, dans le nouvel ordre
            souhaité — doit contenir exactement les mêmes ids que
            GET /checklist-config/items, sans doublon ni omission
            (422 sinon).
          items:
            type: integer
          example: [3, 1, 2]
    ChecklistItemConfig:
      type: object
      properties:
        checklistItemId:
          type: integer
        libelle:
          type: string
          example: Justificatif douanier (D3/BAE) reçu et vérifié
        ordre:
          type: integer
          minimum: 1
        actif:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    ChecklistItemConfigCreateRequest:
      type: object
      required: [libelle]
      properties:
        libelle:
          type: string
          maxLength: 300
    ChecklistItemConfigUpdateRequest:
      type: object
      description: |
        Toutes les propriétés sont optionnelles — seuls les champs
        fournis sont modifiés. ordre n'est PAS modifiable ici (voir
        PATCH /checklist-config/items/reordonner).
      properties:
        libelle:
          type: string
          maxLength: 300
        actif:
          type: boolean
          nullable: true
...
    SignatureVerificationResult:
      type: object
      properties:
        trouve:
          type: boolean
        signatureExistante:
          type: boolean
        nomClient:
          type: string
        typeCompte:
          type: string
          nullable: true
        modeVerification:
          $ref: '#/components/schemas/ModeVerificationSignature'
        # Informations client complémentaires — retournées par ABS2000 au
        # même moment que la vérification de signature (cf. section
        # "Compte client" de CreateDossierRequest), pour pré-remplir le
        # formulaire au lieu d'une ressaisie manuelle. codeBanque est
        # délibérément absent : c'est le code propre à Afriland First
        # Bank CI, pas une donnée client retournée par ABS2000.
        nifClient:
          type: string
          nullable: true
          description: NCC (numéro de compte contribuable) du client.
        adressePostaleClient:
          type: string
          nullable: true
        adresseGeographiqueClient:
          type: string
          nullable: true
        telephoneClient:
          type: string
          nullable: true
        qualiteResidence:
          allOf:
            - $ref: '#/components/schemas/QualiteResidence'
          nullable: true
        dateOuvertureCompte:
          type: string
          format: date
          nullable: true
        anneeExerciceCompte:
          type: integer
          nullable: true
    SoldeClientResult:
      type: object
      properties:
        numCompte:
          type: string
        soldeDisponible:
          type: number
        devise:
          type: string
        suffisant:
          type: boolean
          description: Calculé côté PDOE (soldeDisponible >= montant du dossier), pas renvoyé par ABS2000.
        dateConsultation:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Code erreur métier
          example: SIGNATURE_INVALIDE
          enum:
            - COMPTE_INTROUVABLE
            - SIGNATURE_INVALIDE
            - DOUBLE_PAIEMENT_DETECTE
            - REFERENCE_PAIEMENT_DUPLIQUEE
            - MOTIF_REJET_MANQUANT
            - RESPONSABLE_CORRECTION_MANQUANT
            - DATE_CONFIRMATION_CLIENT_MANQUANTE
            - SOLDE_NON_VERIFIE
            - SOLDE_INSUFFISANT
            - FRACTIONNEMENT_DETECTE
            - PARAMETRAGE_NON_MODIFIABLE
            - VALEUR_HORS_PLAGE
            - PROFIL_INSUFFISANT
            - DOSSIER_INTROUVABLE
            - STATUT_INVALIDE_POUR_ACTION
            - INITIALES_AGENT_MANQUANTES
            - ETAPE_GESTIONNAIRE_DEPASSEE
            - GESTIONNAIRE_LOGIN_MANQUANT
            - LOGIN_AD_DEJA_UTILISE
            - CHAMPS_UTILISATEUR_MANQUANTS
            - UTILISATEUR_INTROUVABLE
            - AUTO_RETRAIT_DROITS_ADMIN_INTERDIT
            - PORTEFEUILLE_DEJA_AFFECTE
            - ETAPE_CODE_DEJA_UTILISE
            - ETAPE_CONFIG_INTROUVABLE
            - ETAPE_DESACTIVATION_REFUSEE
            - ETAPE_ORDRE_INVALIDE
            - MODELE_NOTIFICATION_INTROUVABLE
            - CHECKLIST_ITEM_LIBELLE_MANQUANT
            - CHECKLIST_ITEM_INTROUVABLE
            - CHECKLIST_ORDRE_INVALIDE
            - PARAMETRE_INTROUVABLE
            - FDI_MANQUANTE
            - EXPORT_INTROUVABLE
            - ABS_INDISPONIBLE
            - DOCUMENT_INTROUVABLE
        message:
          type: string
          description: Message lisible par l'humain
        details:
          type: object
          nullable: true
          description: Informations supplémentaires (champs invalides, etc.)